Add  a Client
Sometimes folks may not want to deal with the details and low level
type stuff.  This patch adds a minimal client that can be used as a
receiver to make things a little bit simpler.

If nothing else; it serves as a good example of how to consume the
provider.