Just use standard log package, and add debug opt
Logging is a tricky beast, logging in golang is even more of
a challenge.  With so many options and competing packages for loggers in
golang it's hard to choose what's best.

Let's make an attempt to do things "differently", let's try to keep
things as light-weight and minimal as possible.  This is just a dumb
package after all.

Use that base log package, be minimal in what we log and provide an
option to turn on debug logging.

We'll just use the base log package, remove most of the noise and add
a 'debug' option to the request.  If debug is true we'll print some
extra info to help out with things, but in general we're going to try
and stick to the philosophy of "it either worked, or it returned and
error" and not spam folks with a ton of pointless log messages.