Add the newish generic ListVolumesRequest
We have ListForAccount, ListActive etc, but for a while now SF has
a more generalized ListVolumes API call that let's you just put in all
the various filtering options as parameters (FINALLY!!!!!).

So, add that, frankly the existing methods in this pkg should and will
be deprecated and removed in favor of consolidation and simplification.